rename index files
