+++ src/components/Home/Partners/Partner/Partner.jsx
