+++ src/components/Layout/Layout.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Footer } from '../Home/Footer';
+import { Footer } from '../Home/Footer/Footer';
 
 const Layout = ({ children }) => {
   return (
